requirements.txt
numpy            # Array computations
pytorch          # Tensor computations and Deep Learning
pillow           # Image processing
matplotlib       # Data visualization